requirements.txt
# PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
tqdm>=4.65.0

# Utilities
Pillow>=10.0.0
